Add sufficient error logging support
We do not want to send any fatal errors out to the bot, therefore log
fatal errors to a dedicated log file. It could be the case more than one
error is logged at the same time per app, therefore, add a UUID appId
that'll help with any concurrency issues that may arise.