Introduce SQLite to the project
Our data model evolved to be more relational than originally thought,
and the current DB solution with BoltDB is not satisfying that
requirement. Key-value databases make some of our features require a
lot of code that a relational database could do for us.

SSQLite is an embedable relational database, and will save us from
writing a lot of code to get the same effect from BoltDB.

This commit introduces SQlite to the project, and does not replace the
current functionality.