Set PRAGMA FOREIGN_KEYS to true per repository
SQLite3 does not have FOREIGN KEY constraints enforced by default. We
should establish this once per connection, but we have many functions
that open up a connection to the DB. To not duplicate this code all over
the place, or to run two additional queries per operation in the
repository, just do it once in each repository.