Allow casting 1 vote per suggestion
If someone passes in duplicate suggestion ids, the database will only
insert the last one. This sets that last vote to a preference much
greater than intended. Instead, only allow the first to count, and do
not count any duplicate votes. A user will still be able to cast votes
for multiple suggestions.