Add weekID to suggestion tbl constraint
The primary key on the suggestion table has weekID, author, suggestion.
Added was a unique index just for movieHash for movie lookup. The issue
is that adding the same movie within a different weekID, would result in
a conflict due to the ix_suggestions_movieHash index.

Fix this by adding the weekID to the unique constraint.

Lastly, due to the project restructure, the DB file is now located in
the project root, so update the VS code settings to use the new path.